fix all printing and indent printing as well
