+++ include/ir/value.h
@@ -25,6 +25,7 @@ public:
 
     ASTNS::AST* ast() const override;
     std::string stringify() const override;
+    void definition(std::ostream &os) const;
 
     Type* type() const override;
 
@@ -47,6 +48,7 @@ public:
 
     ASTNS::AST* ast() const override;
     std::string stringify() const override;
+    void definition(std::ostream &os) const;
 
     Type* type() const override;
