+++ src/ir/CMakeLists.txt
@@ -1,1 +1,1 @@
-target_sources(katselc PUBLIC unit.cpp instruction.cpp block.cpp type.cpp builtintype.cpp value.cpp branches.cpp)
+target_sources(katselc PUBLIC unit.cpp instruction.cpp block.cpp type.cpp builtintype.cpp value.cpp branches.cpp function.cpp)
