+++ src/ir/block.cpp
@@ -11,18 +11,29 @@ void Block::add(std::unique_ptr<Instrs::Instruction> instr)
 void Block::branch(std::unique_ptr<Instrs::Br> br)
 {
     if (this->br)
-        reportAbortNoh("Block::branch called multiple time");
+        reportAbortNoh("Block::branch called multiple times");
 
     this->br = std::move(br);
 }
 
 void Block::stringify(std::ostream &os)
 {
-    os << name << "(" << num << "): {\n";
+    os << name << "(" << num << ")";
+}
+
+void Block::definition(std::ostream &os)
+{
+    os << "    " << name << "(" << num << "): {\n";
     for (std::unique_ptr<Instrs::Instruction> const &i : instructions)
+    {
+        os << "        ";
         i->stringify(os);
-    os << "--\n";
+    }
+    os << "    ----\n";
     if (br)
+    {
+        os << "        ";
         br->stringify(os);
-    os << "}\n";
+    }
+    os << "    }\n";
 }
