+++ src/ir/value.cpp
@@ -1,12 +1,11 @@
 #include "ir/value.h"
 #include "message/errors.h"
-#include <sstream>
 
 Register::Register(int index, Type *type, ASTNS::AST *ast, bool temp): temp(temp), index(index), ty(type), _ast(ast) {}
 
 std::string Register::stringify() const
 {
-    return concatMsg(ty->stringify(), " #", index);
+    return concatMsg("#", index);
 }
 ASTNS::AST* Register::ast() const
 {
@@ -21,54 +20,9 @@ bool Register::assignable() const
     return !temp;
 }
 
-Function::Function(FunctionType *ty, std::string name, ASTNS::Function *ast): ty(ty), name(name), _ast(ast), blocki(0), regi(0) {}
-
-void Function::add(std::unique_ptr<Block> block)
+void Register::definition(std::ostream &os) const
 {
-    blocks.push_back(std::move(block));
-}
-
-std::string Function::stringify() const
-{
-    std::stringstream ss;
-    ss << "fun \"" << name << "\" " << ty->stringify() << "\n";
-    ss << "{\n";
-    for (std::unique_ptr<Register> const &r : registers)
-        ss << r->stringify() << std::endl;
-    for (std::unique_ptr<Block> const &b : blocks)
-        b->stringify(ss);
-    ss << "}\n";
-
-    return ss.str();
-}
-ASTNS::AST* Function::ast() const
-{
-    return _ast;
-}
-
-Type* Function::type() const
-{
-    return ty;
-}
-bool Function::assignable() const
-{
-    return false;
-}
-
-Block* Function::addBlock(std::string name)
-{
-    std::unique_ptr<Block> block = std::make_unique<Block>(name, blocki++);
-    Block *blockraw = block.get();
-    blocks.push_back(std::move(block));
-
-    return blockraw;
-}
-Register* Function::addRegister(Type *type, ASTNS::AST *ast, bool temp)
-{
-    std::unique_ptr<Register> reg = std::make_unique<Register>(regi++, type, ast, temp);
-    Register *regraw = reg.get();
-    registers.push_back(std::move(reg));
-    return regraw;
+    os << "    " << ty->stringify() << " #" << index << std::endl;;
 }
 
 ConstInt::ConstInt(BuiltinType *ty, ASTNS::AST *ast, int val): val(val), ty(ty), _ast(ast) {}
